requirements.txt
langchain
flask
pinecone-client
python-dotenv
pypdf
sentence-transformers
ipykernel
pinecone-client
langchain-pinecone